Add memberDetailService for Spring Security
