+++ src/main/java/com/github/springboard/service/MemberService.java
@@ -1,15 +1,11 @@
 package com.github.springboard.service;
 
 import com.github.springboard.domain.Member;
-import com.github.springboard.security.MemberDetails;
 import com.github.springboard.exception.AuthenticationException;
 import com.github.springboard.exception.DuplicateMemberException;
 import com.github.springboard.exception.NotFoundMemberException;
 import com.github.springboard.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +17,7 @@ import java.util.Optional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class MemberService implements UserDetailsService {
+public class MemberService {
 
     private final MemberRepository memberRepository;
 
@@ -81,17 +77,6 @@ public class MemberService implements UserDetailsService {
         if (!members.isEmpty()) {
             throw new DuplicateMemberException("이미 존재하는 아이디입니다.");
         }
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        List<Member> members = memberRepository.findByUsername(username);
-        if (members.isEmpty()) {
-            throw new NotFoundMemberException("존재하지 않는 회원입니다.");
-        }
-
-        Member member = members.get(0);
-        return new MemberDetails(member);
     }
 
 }
